requirements.txt
mlflow==2.3
Pillow==9.5.0
Django==4.2
psycopg2==2.9.6
bcrypt
nodeenv
django-cors-headers
djangorestframework
django-filter
pandas
numpy
seaborn
plotly==5.15.0
scikit-learn
tensorflow
markdown